pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "control"
dynamic = ["version"]
license = "Apache-2.0"
authors = [{ name = "Newton Developers", email = "warp-python@nvidia.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: GPU :: NVIDIA CUDA",
]
description = "Actuator-based control library for GPU-accelerated physics simulations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["warp-lang>=1.11.0.dev20251205"]

[tool.hatch.version]
path = "control/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["control"]

[[tool.uv.index]]
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "nvidia"
url = "https://pypi.nvidia.com/"
explicit = true

[tool.uv.sources]
warp-lang = { index = "nvidia" }

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "I", "F", "W", "B", "UP"]
ignore = ["E501", "E741", "UP018", "F811"]

[tool.ruff.format]
quote-style = "double"
